third_party/WebKit/LayoutTests/editing/selection/mouse/drag_selection_update_crash.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<div style="height:1000px"><span contenteditable id="target">abc&#x5D0;&#x5D1;&#x5D2;</span></div>
<div id="log"></div>
<script>
test(function(){
    assert_true(eventSender != null, 'eventSender is required');
    var y = target.offsetTop + target.offsetHeight / 2;
    eventSender.dragMode = false;
    eventSender.mouseMoveTo(target.offsetLeft + target.offsetWidth, y);
    eventSender.mouseDown();
    eventSender.mouseMoveTo(target.offsetLeft + target.offsetWidth - 10, y);
});
</script>
